Restructure topic-details module
