+++ src/app/topic-details/persisted-items/new-persisted-item/new-persisted-item.component.css
