+++ src/app/topic-details/persisted-items/new-persisted-item/new-persisted-item.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
-import {NavigationService} from '../../core/navigation.service';
 import {PersistedItemsService, PublishItemErrors} from '../persisted-items.service';
 import {NgForm} from '@angular/forms';
+import {NavigationService} from '../../../core/navigation.service';
 
 @Component({
   selector: 'xgb-new-persisted-item',
