+++ src/app/topic-details/persisted-items/persisted-items.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {IqType, XmppService} from '../core/xmpp/xmpp.service';
+import {IqType, XmppService} from '../../core/xmpp/xmpp.service';
 
 export class PersistedItem {
   /**
