+++ src/app/topic-details/persisted-items/persisted-items/persisted-items.component.css
