+++ src/app/topic-details/topic-details-router.module.ts
@@ -1,13 +1,13 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {TopicDetailsComponent} from './topic-details/topic-details.component';
+import {TopicDetailsComponent} from './topic-details.component';
 import {TopicDetailsConfigComponent} from './topic-details-config/topic-details-config.component';
 import {TopicAffiliationsComponent} from './topic-affiliations/topic-affiliations.component';
 import {XmppFeatureGuardService as XmppFeatureGuard} from '../core/xmpp/xmpp-feature-guard.service';
-import {PersistedItemsComponent} from './persisted-items/persisted-items.component';
-import {NewPersistedItemComponent} from './new-persisted-item/new-persisted-item.component';
-import {TopicSubscriptionComponent} from './topic-subscription/topic-subscription.component';
-import {NewTopicSubscriptionComponent} from './new-topic-subscription/new-topic-subscription.component';
+import {PersistedItemsComponent} from './persisted-items/persisted-items/persisted-items.component';
+import {NewPersistedItemComponent} from './persisted-items/new-persisted-item/new-persisted-item.component';
+import {TopicSubscriptionComponent} from './topic-subscription/topic-subscription/topic-subscription.component';
+import {NewTopicSubscriptionComponent} from './topic-subscription/new-topic-subscription/new-topic-subscription.component';
 
 
 const routes: Routes = [
