+++ src/app/topic-details/topic-details.component.ts
