+++ src/app/topic-details/topic-details.module.ts
@@ -6,14 +6,14 @@ import {TopicDetailsConfigComponent} from './topic-details-config/topic-details-
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {TopicDetailsService} from './topic-details.service';
 import {TopicWidgetsModule} from '../topic-widgets/topic-widgets.module';
-import {TopicDetailsComponent} from './topic-details/topic-details.component';
+import {TopicDetailsComponent} from './topic-details.component';
 import {TopicAffiliationsComponent} from './topic-affiliations/topic-affiliations.component';
-import {PersistedItemsComponent} from './persisted-items/persisted-items.component';
-import {PersistedItemsService} from './persisted-items.service';
-import {NewPersistedItemComponent} from './new-persisted-item/new-persisted-item.component';
-import {TopicSubscriptionComponent} from './topic-subscription/topic-subscription.component';
-import {TopicSubscriptionService} from './topic-subscription.service';
-import {NewTopicSubscriptionComponent} from './new-topic-subscription/new-topic-subscription.component';
+import {PersistedItemsComponent} from './persisted-items/persisted-items/persisted-items.component';
+import {PersistedItemsService} from './persisted-items/persisted-items.service';
+import {NewPersistedItemComponent} from './persisted-items/new-persisted-item/new-persisted-item.component';
+import {TopicSubscriptionComponent} from './topic-subscription/topic-subscription/topic-subscription.component';
+import {TopicSubscriptionService} from './topic-subscription/topic-subscription.service';
+import {NewTopicSubscriptionComponent} from './topic-subscription/new-topic-subscription/new-topic-subscription.component';
 
 
 @NgModule({
